feat(QuestionExplanation) add explanation question whit conditional buttons and refactor interfaces
