+++ src/components/atoms/RadioInput.tsx
@@ -19,7 +19,6 @@ export const RadioInput: React.FC<RadioButtonProps> = ({
         checked={selectedOption === option}
         className="form-radio"
       />
-      {/*     <span>{option}</span> */}
     </label>
   );
 };
